+++ app/src/main/java/woowacourse/payments/ui/component/CardInputFields.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.ui.registercard.CardInputFieldStateHolder
+import woowacourse.payments.ui.CardInputFieldStateHolder
 
 @Composable
 fun CardInputFields(
